sniugb-backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.2
APScheduler==3.10.4
bleach==6.1.0
requests==2.32.3
pydantic==2.9.2
structlog==24.1.0
sentry-sdk==2.14.0
prometheus-fastapi-instrumentator==7.0.0
slowapi==0.1.9
pytest==8.3.3
ruff==0.7.0
mypy==1.11.2
pydantic[email]
sendgrid
python-multipart
pandas
pillow
aiofiles
python-slugify
